+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Basic CRUD for backoffice Research areas (@mkasztelnik)
 - Basic CRUD for backoffice Categories (@mkasztelnik)
 - New custom fields for voucher in jira (@martaswiatkowska)
+- Basic CRUD for backoffice Providers (@mkasztelnik)
 
 ### Changed